requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-bootstrap3==14.1.0
gunicorn==20.0.4
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.10
psycopg2>=2.6.1
